refactor(markets): reorg file structure
