+++ markets/retrievaladapter/client.go
@@ -2,13 +2,12 @@ package retrievaladapter
 
 import (
 	"context"
-	"github.com/filecoin-project/lotus/lib/sharedutils"
 
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-fil-markets/retrievalmarket"
 	retrievaltoken "github.com/filecoin-project/go-fil-markets/shared/tokenamount"
 	retrievaltypes "github.com/filecoin-project/go-fil-markets/shared/types"
-
+	"github.com/filecoin-project/lotus/markets/utils"
 	payapi "github.com/filecoin-project/lotus/node/impl/paych"
 	"github.com/filecoin-project/lotus/paych"
 )
@@ -27,7 +26,7 @@ func NewRetrievalClientNode(pmgr *paych.Manager, payapi payapi.PaychAPI) retriev
 // GetOrCreatePaymentChannel sets up a new payment channel if one does not exist
 // between a client and a miner and insures the client has the given amount of funds available in the channel
 func (rcn *retrievalClientNode) GetOrCreatePaymentChannel(ctx context.Context, clientAddress address.Address, minerAddress address.Address, clientFundsAvailable retrievaltoken.TokenAmount) (address.Address, error) {
-	paych, _, err := rcn.pmgr.GetPaych(ctx, clientAddress, minerAddress, sharedutils.FromSharedTokenAmount(clientFundsAvailable))
+	paych, _, err := rcn.pmgr.GetPaych(ctx, clientAddress, minerAddress, utils.FromSharedTokenAmount(clientFundsAvailable))
 	return paych, err
 }
 
@@ -42,9 +41,9 @@ func (rcn *retrievalClientNode) AllocateLane(paymentChannel address.Address) (ui
 // given payment channel so that all the payment vouchers in the lane add up
 // to the given amount (so the payment voucher will be for the difference)
 func (rcn *retrievalClientNode) CreatePaymentVoucher(ctx context.Context, paymentChannel address.Address, amount retrievaltoken.TokenAmount, lane uint64) (*retrievaltypes.SignedVoucher, error) {
-	voucher, err := rcn.payapi.PaychVoucherCreate(ctx, paymentChannel, sharedutils.FromSharedTokenAmount(amount), lane)
+	voucher, err := rcn.payapi.PaychVoucherCreate(ctx, paymentChannel, utils.FromSharedTokenAmount(amount), lane)
 	if err != nil {
 		return nil, err
 	}
-	return sharedutils.ToSharedSignedVoucher(voucher)
+	return utils.ToSharedSignedVoucher(voucher)
 }
