+++ markets/retrievaladapter/provider.go
@@ -2,16 +2,17 @@ package retrievaladapter
 
 import (
 	"context"
-	"github.com/filecoin-project/lotus/lib/sharedutils"
+
+	"github.com/ipfs/go-cid"
+	blockstore "github.com/ipfs/go-ipfs-blockstore"
 
 	"github.com/filecoin-project/go-address"
 	"github.com/filecoin-project/go-fil-markets/retrievalmarket"
 	retrievaltoken "github.com/filecoin-project/go-fil-markets/shared/tokenamount"
 	retrievaltypes "github.com/filecoin-project/go-fil-markets/shared/types"
 	"github.com/filecoin-project/lotus/api"
+	"github.com/filecoin-project/lotus/markets/utils"
 	"github.com/filecoin-project/lotus/storage/sectorblocks"
-	"github.com/ipfs/go-cid"
-	blockstore "github.com/ipfs/go-ipfs-blockstore"
 )
 
 type retrievalProviderNode struct {
@@ -38,10 +39,10 @@ func (rpn *retrievalProviderNode) SealedBlockstore(approveUnseal func() error) b
 }
 
 func (rpn *retrievalProviderNode) SavePaymentVoucher(ctx context.Context, paymentChannel address.Address, voucher *retrievaltypes.SignedVoucher, proof []byte, expectedAmount retrievaltoken.TokenAmount) (retrievaltoken.TokenAmount, error) {
-	localVoucher, err := sharedutils.FromSharedSignedVoucher(voucher)
+	localVoucher, err := utils.FromSharedSignedVoucher(voucher)
 	if err != nil {
 		return retrievaltoken.FromInt(0), err
 	}
-	added, err := rpn.full.PaychVoucherAdd(ctx, paymentChannel, localVoucher, proof, sharedutils.FromSharedTokenAmount(expectedAmount))
-	return sharedutils.ToSharedTokenAmount(added), err
+	added, err := rpn.full.PaychVoucherAdd(ctx, paymentChannel, localVoucher, proof, utils.FromSharedTokenAmount(expectedAmount))
+	return utils.ToSharedTokenAmount(added), err
 }
